refactor: Complete migration to C++ thread registry implementation
- Remove C implementation of thread_registry.c entirely
- Remove registry_selector runtime selection mechanism
- Remove feature flags and conditional compilation
- Convert all tests to use C++ API directly
- Fix memory layout to use pointer-based allocation instead of inline buffers
- Change thread_id type from uint32_t to uintptr_t for platform portability
- Fix SPSC queue wraparound test to use correct ring indices

The C++ implementation is now the sole thread registry with:
- Full lock-free SPSC queue functionality
- Cache-line aligned structures for performance
- Platform-portable thread IDs
- Proper memory management with separate ring buffer allocation
- All unit tests passing (12/12)

Note: Integration test pending conversion in next iteration

BREAKING CHANGE: Removes C thread registry API compatibility

+++ tracer_backend/include/thread_registry.h
@@ -37,7 +37,7 @@ ThreadRegistry* thread_registry_attach(void* memory);
 // thread_id: system thread ID for this thread
 // Returns: pointer to ThreadLaneSet for this thread, or NULL if full
 // Memory ordering: Uses memory_order_acq_rel for slot allocation
-ThreadLaneSet* thread_registry_register(ThreadRegistry* registry, uint32_t thread_id);
+ThreadLaneSet* thread_registry_register(ThreadRegistry* registry, uintptr_t thread_id);
 
 extern bool needs_log_thread_registry_registry;
